Move `value_getter` into the utils folder
Since it is templated, `value_getter` does not rely on any one
`visitor_base`.
